requirements.txt
youtube-dl
mutagen
requests
python-telegram-bot
pymongo[srv]